Move program source code to src directory
